# Fix: up to 80% speed and memory optimizations for slices