@samber 2022-09-15
Oops, thanks for these obvious optimization!

![](https://media3.giphy.com/media/Hpj2jSuJNLe05zezy9/giphy.gif?cid=ecf05e4745lnujczeknupcg1f6z5qxdzhijyv8o74lzu4fx9&rid=giphy.gif&ct=g)